badgeware: refactor for new badge API.
