+++ modules/common/badgeware/filesystem.py
@@ -1,0 +1,22 @@
+import builtins
+import os
+
+
+def file_exists(path):
+    try:
+        os.stat(path)
+        return True
+    except OSError:
+        return False
+
+
+def is_dir(path):
+    try:
+        flags = os.stat(path)
+        return flags[0] & 0x4000  # is a directory
+    except:  # noqa: E722
+        return False
+
+
+builtins.file_exists = file_exists
+builtins.is_dir = is_dir
