+++ modules/common/badgeware/sprite.py
@@ -1,0 +1,50 @@
+import builtins
+
+
+class SpriteSheet:
+    def __init__(self, file, columns, rows):
+        self.image = image.load(file)
+        self.sw = int(self.image.width / columns)
+        self.sh = int(self.image.height / rows)
+
+        self.sprites = []
+        for x in range(columns):
+            column = []
+            for y in range(rows):
+                sprite = self.image.window(self.sw * x, self.sh * y, self.sw, self.sh)
+                column.append(sprite)
+            self.sprites.append(column)
+
+    def sprite(self, x, y):
+        return self.sprites[x][y]
+
+    def animation(self, x=0, y=0, count=None, horizontal=True):
+        if not count:
+            count = int(self.image.width / self.sw)
+        return AnimatedSprite(self, x, y, count, horizontal)
+
+
+class AnimatedSprite:
+    def __init__(self, spritesheet, x, y, count, horizontal=True):
+        self.spritesheet = spritesheet
+        self.frames = []
+        for _ in range(count):
+            self.frames.append((x, y))
+            if horizontal:
+                x += 1
+            else:
+                y += 1
+
+    def frame(self, frame_index=0):
+        frame_index = int(frame_index)
+        frame_index %= len(self.frames)
+        return self.spritesheet.sprite(
+            self.frames[frame_index][0], self.frames[frame_index][1]
+        )
+
+    def count(self):
+        return len(self.frames)
+
+
+builtins.SpriteSheet = SpriteSheet
+builtins.AnimatedSprite = AnimatedSprite
